refactor:[JIRATASK-000] Typography Component
