+++ src/components/Typography/Typography.module.css
@@ -1,0 +1,1 @@
+@import '_colors';
